+++ codegen/generate.go
@@ -8,10 +8,9 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/vektah/gqlparser/v2/ast"
-
 	"github.com/99designs/gqlgen/codegen/config"
 	"github.com/99designs/gqlgen/codegen/templates"
+	"github.com/vektah/gqlparser/v2/ast"
 )
 
 func GenerateCode(data *Data) error {
